wagmi: import rainbowkit and walletconect only on web
