+++ apps/next/pages/_app.tsx
@@ -6,16 +6,14 @@ import 'raf/polyfill'
 import '@my/ui/src/config/fonts.css'
 
 import { type ColorScheme, NextThemeProvider, useRootTheme } from '@tamagui/next-theme'
-import { Provider } from 'app/provider/index.web'
+
 import type { AuthProviderProps } from 'app/provider/auth'
 import { api } from 'app/utils/api'
 import type { NextPage } from 'next'
 import Head from 'next/head'
-import { RainbowKitProvider } from '@rainbow-me/rainbowkit'
-
 import type { ReactElement, ReactNode } from 'react'
 import type { SolitoAppProps } from 'solito'
-import { baseMainnetClient } from 'app/utils/viem'
+import { Provider } from 'app/provider'
 
 if (process.env.NODE_ENV === 'production') {
   require('../public/tamagui.css')
@@ -62,9 +60,7 @@ function MyApp({
         }}
       >
         <Provider initialSession={pageProps.initialSession}>
-          <RainbowKitProvider initialChain={baseMainnetClient.chain}>
-            {getLayout(<Component {...pageProps} />)}
-          </RainbowKitProvider>
+          {getLayout(<Component {...pageProps} />)}
         </Provider>
       </NextThemeProvider>
     </>
