+++ packages/app/provider/rainbowkit/RainbowkitProvider.native.tsx
@@ -1,0 +1,4 @@
+// noop on native for now...
+export function RainbowkitProvider({ children }: { children: React.ReactNode }) {
+  return <>{children}</>
+}
