+++ packages/app/provider/rainbowkit/RainbowkitProvider.tsx
@@ -1,0 +1,6 @@
+import { RainbowKitProvider } from '@rainbow-me/rainbowkit'
+import { baseMainnetClient } from 'app/utils/viem'
+
+export function RainbowkitProvider({ children }: { children: React.ReactNode }) {
+  return <RainbowKitProvider initialChain={baseMainnetClient.chain}>{children}</RainbowKitProvider>
+}
