+++ packages/app/provider/rainbowkit/index.ts
@@ -1,0 +1,1 @@
+export * from './RainbowkitProvider'
